refactor, fix mixes, make cutomizable
